Massive refactoring
Moved pinchy to pinchyd, now a dbus service. Added pinchy, a client for
that service.

Some improvements to printing of ppoll ended up tangled into this
somehow and I'm right now too lazy to fix it. Since it was already
very noisy I decided to add some more to protect from noise in the
future: Cargo.lock got added to gitignore.

The dbus service is just a shell for now, it doesn't really work with
the tracing yet. It just allows the client to request tracing for a
PID and shares a file descriptor that we will later use to pass trace
messages to the client.