Allow registering PID/fd pairs for tracing
This provides the missing functionality to the trace_pid DBus method
that can now set up the forwarding of trace messages through the shared
fd.

Keep in mind this is a massive security hazard if used as is, as there
is no authorization in place. If the server is running, anyone can
call this method and get tracing data for any process in the system.