Move event writer into is own task
This makes the code a bit easier to follow, as the writing to the subs
is not hidden inside event::handle_event().

Most importantly, it also guarantees ordering. With each writer being
spawned independently there is always a chance a bunch will be queued
up and resolved out of order, which caused me to see a read() produce
EOF followed by a successful reading of bytes ;P