Refactor try_* into inner functions
The x() calls try_x() pattern used by the aya template is great because
it allows for ergonomic usage of Result<>. However, the verifier gets
angry about the functions that anyone could in theory call potentially
exceeding the maximum stack size.

We move the try_x() functions inside with the name inner(), thus hiding
them from the wrath of the verifier - who is now satisfied, as there
are programs that can be tail called, not regular callables.