Basic testing
Initial testing harness and a couple simple tests. For now, this needs
to run as root. Ideally this will work with something like virtme-ng,
so that it can run even in CI, but more work is needed.

To run it right now you need something like this:

> cargo --config "target.'cfg(all())'.runner=['/bin/sudo', '-s']" test

This makes the build happen with your regular user, but uses sudo to
run the generated binary. You also need to have the dbus-launch tool
installed.