Pinchy client can now start a subprocess and trace it
AI was only a bit helpful for this one, I'm afraid. And std::process
was not at all. To properly trace, we need to be able to ensure the
tracing is set up after we know the PID, but before the child process
is replaced by the one we want to trace.

There is no good way of doing that right now with std::process::Command
so we get dirty with unsafe { libc } all over the place.

This change also includes a new test helper that will be used to
generate syscall workloads for our tracing tests, as well as our
first pinchy client test. Since we'll need some data for reading in
many tests, I also added a copy of the GPLv2.

Finally, this change also uses moves the dbus-launch usage from the
launch of pinchyd to a OnceCell, as we now need to run multiple
processes in the same bus for testing.