Add support for the execve syscall (assisted GPT 4.1)
This is a tricky one, as the process memory goes away after it is done,
so we need to add an enter handler specifically for execve, so that we
can read the arguments and save them for the exit one.

AI was surprisingly useful, doing most of the work. It got a bit confused
when dealing with the string and array truncation, but otherwise I am
amazed by how much it was able to do.