Move pretty printing to the client (assisted)
This change basically moves the events.rs file from server to client.
Server now simply forwards SyscallEvent structs through the wire to
the client, which reads and parses them in order.

It also removes the ability to give the server a PID to trace and
print itself, as that is not really possible anymore - we could
keep that ability, but I think we have enough infrastructure to rely
solely on the client at this point. This also means the basic_reads
test is obsolete, so remove.

I initially thought about making the server provide nicely formatted
output that can be reused by various types of clients. The more I
think about that, however, the more I think the clients would actually
prefer the raw data to parse as they want.

Also, I can break apart event parsing into a library with reusable
parts when I get there (I have been pondering reimplementing lupa
as a pinchyd client).

The other thing I did not want was to add a bunch of serialization
overhead right now. I decided I should just, for now, assume server
and client are simply built from the same codebase and use the same
types. If I eventually need to make the data more portable I can write
a client whose job is to do that.

I feel like I was prematurely generalizing and causing more code to
be run in a privileged environment unnecessarily. This fixes that
mistake.