Report lost events, default to warn log level
The end game will probably be switching to a different map type, like
a ring buffer that can retry and slow down the traced program. But for
now let's just keep track of dropped events and work towards reducing
that number, so we can have as low overhead as possible.