Add support for prctl
This is a bit of a special one, as we are using the generic eBPF
handler, as we want all of the data and do no reading from user
memory as of yet.

For now we only decode operation and show the appropriate number of
arguments for a given operation as hexadecimal values. In the
future we may do further decoding, which will require a specific
eBPF-side handler.