Major formatting rework - begin
We add a new formatting framework that allows us to decouple formatting
from the parsing / structuring of the output. Instead of a format, we
have functions to add arguments, arrays, structures, which also encodes
the idea of depth - that can be used by the formatter to indent, for
instance - and possibly more advanced features like coloring also become
possible.

This change adds tests for one-line formatting of all of the syscalls
we support, making sure the output is as expected and that the new
model exactly duplicates the manually formatted versions.