Rework how writer reaping works
Instead of doing reaping in the idle quit loop, do it on a timeout along
with the writer, making it easier to share a big chunk of the code.

Also, since flushing is apparently not enough to detect the other end
being closed, perform a manual poll of the file descriptor after
flushing to identify closed pipes.