Begin rework of ringbuf handling
Up to this point, we dealt with all of the data related to a syscall
on the stack first to then create the ringbuf entry and submit.

With this change, we begin implementation of a new strategy: reserve
the ringbuf memory at the start of the tracepoint handler and use that
memory directly as much as possible when reading data from memory.

Ideally this will allow us to increase the amount of data we read from
syscalls, which is getting more and more tricky to deal with on the
stack.