Use a fixed nightly version for eBPF target
This is a feature that has recently been introduced to arya-build but is
still not present on a release, so we depend on the git repository for
now.

I have been debugging a verifier failure that was only happening in CI
and have come to the conclusion it is a codegen issue, likely caused by
a regression in LLVM (its eBPF target, at least). I could reproduce on
my machine after updating my nightly.

Using a fixed version should avoid that kind of thing happening out of
the blue, and will make the CI more robust.